:boom: Changes on status & Architecture
Remove `cmd->status`, move functions in betters files
