+++ minishell/Sources/execution/main_execute.c
@@ -6,18 +6,18 @@
 /*   By: tonted <tonted@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/26 21:35:20 by tonted            #+#    #+#             */
-/*   Updated: 2022/10/14 02:39:20 by tonted           ###   ########.fr       */
+/*   Updated: 2022/10/14 21:17:36 by tonted           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-// TODO refactor `get_status()`
+// TODO refactor `get_value_status()`
 static bool	_continue(t_cmd *cmd, char ***envp, int ctrl)
 {
 	return (ft_strlen(*cmd->cmd) && !exec_builtins(cmd, envp, MAIN)
 		&& ((ctrl != AND && ctrl != OR)
-			|| ((ctrl == AND && !get_status()) || (ctrl == OR && get_status()))));
+			|| ((ctrl == AND && !get_value_status()) || (ctrl == OR && get_value_status()))));
 }
 
 //TODO explications sur le ctrl?
@@ -45,5 +45,5 @@ int	execute(char *command, char ***envp)
 	if (r_cmd)
 		cmd = r_cmd;
 	free_cmd(cmd);
-	return (get_status());
+	return (get_value_status());
 }
